refactor(s2-automate): move ErrorCategory to s2-automate-dsl
Lower the retry/remediation taxonomy from s2-automate-core
(s2.automate.core.persist) to s2-automate-dsl (s2.dsl.automate) so
downstream layers — notably chaincode-api-fabric, which carries TxOutcome
— can share the same enum without depending on the whole engine.

Updates the same-package consumers (PersistOutcome, LoadOutcome,
AutomatePersistFailure) to add an explicit DSL import, and adjusts the 4
commonTest files that referenced the old package-local symbol.

No semantic change; the enum's @JsExport / @JsName contract is preserved
so the TypeScript surface generated from the DSL module is unchanged.
