+++ s2-automate/s2-automate-core/src/commonMain/kotlin/s2/automate/core/persist/PersistOutcome.kt
@@ -1,5 +1,6 @@
 package s2.automate.core.persist
 
+import s2.dsl.automate.ErrorCategory
 import s2.dsl.automate.S2Error
 
 sealed interface PersistOutcome<EVENT> {
