+++ s2-automate/s2-automate-core/src/commonTest/kotlin/s2/automate/core/storing/S2AutomateStoringEvolverImplFailurePublishTest.kt
@@ -15,8 +15,8 @@ import s2.dsl.automate.S2Automate
 import s2.automate.core.engine.S2AutomateEngine
 import s2.automate.core.engine.S2AutomateOutcomeEngine
 import s2.automate.core.persist.AutomatePersistFailure
-import s2.automate.core.persist.ErrorCategory
 import s2.automate.core.persist.PersistOutcome
+import s2.dsl.automate.ErrorCategory
 import s2.dsl.automate.Evt
 import s2.dsl.automate.S2Command
 import s2.dsl.automate.S2InitCommand
